# api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Flowfish Platform API
  description: |
    # Flowfish Platform API
    
    REST API documentation for the eBPF-based Kubernetes/OpenShift application communication and dependency mapping platform.
    
    ## Key Features
    
    - **Automatic Discovery** — Automatically detect inter-pod communications using eBPF
    - **Real-Time Dependency Map** — Live visualization of service dependency graphs
    - **Anomaly Detection** — Intelligent anomaly detection powered by statistical analysis
    - **Multi-Cluster** — Manage multiple clusters from a single interface
    - **Integration** — Dependency intelligence APIs for CI/CD pipeline integration
    - **Pod & Deployment Annotations** — Rich metadata enrichment including annotations merged from owning Deployments/StatefulSets
    - **Import/Export** — Data import and export in multiple formats
    
    ## Authentication
    
    The API supports two authentication methods:
    
    1. **JWT (JSON Web Token)** — Send a POST request to `/api/v1/auth/login` with username and password, then include the returned `access_token` in subsequent requests via `Authorization: Bearer <token>` header.
    2. **API Key** — Generate an API key from Settings > API Keys, then include it via `X-API-Key: <key>` header. Recommended for CI/CD pipelines and AI agent integrations.
    
    ## Rate Limiting
    
    - **Authenticated Users**: 1,000 requests/hour
    - **Unauthenticated**: 100 requests/hour
    
...
  contact:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authorization header using the Bearer scheme
  schemas:
...
    Communication:
      type: object
      properties:
        id:
          type: integer
        cluster_id:
          type: integer
        source_namespace:
          type: string
        source_workload:
          type: string
        source_workload_type:
          type: string
        destination_namespace:
          type: string
        destination_workload:
          type: string
        destination_workload_type:
          type: string
        destination_ip:
          type: string
        destination_port:
          type: integer
        protocol:
          type: string
        first_seen:
          type: string
          format: date-time
        last_seen:
          type: string
          format: date-time
        request_count:
          type: integer
        request_rate_per_second:
          type: number
          format: float
        avg_latency_ms:
          type: number
          format: float
        risk_score:
          type: integer
          minimum: 0
          maximum: 100
        risk_level:
          type: string
          enum: [low, medium, high, critical]
        is_cross_namespace:
          type: boolean
        is_external:
          type: boolean
        is_active:
          type: boolean
...
    DependencySummaryService:
      type: object
      properties:
        name:
          type: string
        namespace:
          type: string
        kind:
          type: string
        annotations:
          type: object
          additionalProperties:
            type: string
        labels:
          type: object
          additionalProperties:
            type: string
        is_critical:
          type: boolean
        service_type:
          type: string
        port:
          type: integer
        hop_count:
          type: integer
          description: Number of hops from the upstream service (1 = direct neighbor). Present when depth > 1.
    DependencySummaryGroup:
      type: object
      properties:
        total:
          type: integer
        critical_count:
          type: integer
        by_category:
          type: object
          additionalProperties:
            type: array
            items:
              $ref: '#/components/schemas/DependencySummaryService'
    DependencySummary:
      type: object
      properties:
        total_matched:
          type: integer
          description: Number of unique upstream workloads matched
        total_downstream_unique:
          type: integer
          description: Unique downstream dependency count across all matched services
        total_callers_unique:
          type: integer
          description: Unique caller count across all matched services
        downstream_critical_count:
          type: integer
          description: Critical downstream dependency count (globally deduplicated)
        callers_critical_count:
          type: integer
          description: Critical caller count (globally deduplicated)
    MatchedService:
      type: object
      properties:
        name:
          type: string
        namespace:
          type: string
        kind:
          type: string
        annotations:
          type: object
          additionalProperties:
            type: string
        labels:
          type: object
          additionalProperties:
            type: string
        downstream:
          $ref: '#/components/schemas/DependencySummaryGroup'
        callers:
          $ref: '#/components/schemas/DependencySummaryGroup'
    DependencySummaryResponse:
      type: object
      properties:
        success:
          type: boolean
        analysis_ids:
          type: array
          items:
            type: integer
        multi_service:
          type: boolean
          description: True when multiple upstream workloads matched
        summary:
          $ref: '#/components/schemas/DependencySummary'
        service:
          $ref: '#/components/schemas/DependencySummaryService'
          description: Aggregate service label. For single match shows actual service info; for multi-match shows summary label.
        matched_services:
          type: array
          description: Each matched upstream workload with its own downstream/callers breakdown. Always present when success is true.
          items:
            $ref: '#/components/schemas/MatchedService'
        error:
          type: string
    GraphData:
      type: object
      properties:
        nodes:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              type:
                type: string
              name:
                type: string
              namespace:
                type: string
              labels:
                type: object
                additionalProperties:
                  type: string
              annotations:
                type: object
                additionalProperties:
                  type: string
              metadata:
                type: object
        edges:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              source:
                type: string
              target:
                type: string
              port:
                type: integer
              protocol:
                type: string
              metrics:
                type: object
security:
  - BearerAuth: []
paths:
...
  /communications:
    get:
      tags:
        - Communications
      summary: List communication records
      parameters:
        - name: cluster_id
          in: query
          required: true
          schema:
            type: integer
        - name: namespace
          in: query
          schema:
            type: string
        - name: source_workload
          in: query
          schema:
            type: string
        - name: destination_workload
          in: query
          schema:
            type: string
        - name: risk_level
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: is_active
          in: query
          schema:
            type: boolean
            default: true
        - name: from_time
          in: query
          schema:
            type: string
            format: date-time
        - name: to_time
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Communication records
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Communication'
                  total_count:
                    type: integer
  /dependencies/graph:
    get:
      tags:
        - Dependencies
      summary: Get dependency map
      description: Returns the dependency map as a graph with nodes and edges
      parameters:
        - name: cluster_id
          in: query
          description: Optional; used for filtering in multi-cluster scenarios
          schema:
            type: integer
        - name: analysis_id
          in: query
          schema:
            type: integer
        - name: namespace
          in: query
          schema:
            type: string
        - name: depth
          in: query
          schema:
            type: integer
            default: 2
        - name: search
          in: query
          schema:
            type: string
            minLength: 3
      responses:
        '200':
          description: Graph data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphData'
...
  /communications/dependencies/summary:
    get:
      tags:
        - Integration
      summary: Per-service dependency summary grouped by category for CI/CD and AI integration
      description: |
        Returns a per-service dependency breakdown for CI/CD pipelines, AI agents, and automation.
        Each matched upstream workload includes its own downstream and callers grouped by service_category.
        A top-level summary provides globally deduplicated aggregate counts.
        Dependencies include hop_count when depth > 1 to distinguish direct from indirect paths.
        Annotations and labels are exposed for cross-project impact analysis (e.g. git-repo URLs, team info).
        Requires at least one analysis_id and one search parameter.
      parameters:
        - name: analysis_ids
          in: query
          required: true
          description: Analysis IDs (at least one required)
          schema:
            type: array
            items:
              type: integer
            minItems: 1
        - name: cluster_id
          in: query
          schema:
            type: integer
        - name: pod_name
          in: query
          schema:
            type: string
        - name: namespace
          in: query
          schema:
            type: string
        - name: owner_name
          in: query
          schema:
            type: string
        - name: annotation_key
          in: query
          schema:
            type: string
        - name: annotation_value
          in: query
          schema:
            type: string
        - name: label_key
          in: query
          schema:
            type: string
        - name: label_value
          in: query
          schema:
            type: string
        - name: ip
          in: query
          schema:
            type: string
        - name: depth
          in: query
          description: |
            Graph traversal depth (1-5 hops). Default 1 returns direct neighbors only.
            Higher values discover indirect (transitive) dependencies. Each dependency's hop_count
            field indicates how many hops it is from the upstream service.
          schema:
            type: integer
            minimum: 1
            maximum: 5
            default: 1
      responses:
        '200':
          description: Per-service dependency breakdown with aggregate summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DependencySummaryResponse'
...
  /l7/communications:
    get:
      tags: [L7 Communications]
      summary: List L7 workload-to-workload communications
      parameters:
        - name: analysis_id
          in: query
          required: true
          schema: { type: integer }
        - name: cluster_id
          in: query
          schema: { type: integer }
        - name: namespace
          in: query
          schema: { type: string }
        - name: protocol
          in: query
          schema: { type: string, enum: [http, grpc, dns] }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
      responses:
        '200':
          description: L7 communication list
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /l7/dependencies/graph:
    get:
      tags: [L7 Communications]
      summary: L7 dependency graph (nodes and edges)
      parameters:
        - name: analysis_id
          in: query
          required: true
          schema: { type: integer }
        - name: cluster_id
          in: query
          schema: { type: integer }
        - name: namespace
          in: query
          schema: { type: string }
        - name: protocol
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Graph with L7Workload nodes and L7_COMMUNICATES_WITH edges
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /l7/communications/stats:
    get:
      tags: [L7 Communications]
      summary: Aggregated L7 communication statistics
      parameters:
        - name: analysis_id
          in: query
          required: true
          schema: { type: integer }
        - name: cluster_id
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: Stats including total workloads, edges, requests, errors, avg latency
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /l7/communications/error-stats:
    get:
      tags: [L7 Communications]
      summary: L7 error totals and breakdown by protocol
      parameters:
        - name: analysis_id
          in: query
          required: true
          schema: { type: integer }
        - name: cluster_id
          in: query
          schema: { type: integer }
        - name: namespace
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Error stats per protocol
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /l7/dependencies/summary:
    get:
      tags: [L7 Communications]
      summary: Per-workload L7 dependency summary
      description: |
        Returns one entry per L7 workload (Beyla-observed application identity)
        with aggregated inbound/outbound edge counts, request totals, error totals,
        and error rate.

        When any of `annotation_key`, `annotation_value`, `label_key`,
        `label_value`, `owner_name`, `pod_name`, or `workload_name` is supplied,
        the response narrows to matched workloads + their immediate neighbours
        and each workload carries an `is_matched` flag. Glob patterns
        (`*`, `?`, `[seq]`) are supported on annotation/label key and value
        fields. `owner_name` is an L4-compatibility alias for `workload_name`.
      parameters:
        - name: analysis_id
          in: query
          required: true
          description: Analysis ID; accepts the parent analysis ID or a multi-cluster sub-analysis ID
          schema: { type: string }
        - name: cluster_id
          in: query
          schema: { type: string }
        - name: namespace
          in: query
          schema: { type: string }
        - name: include_metadata
          in: query
          description: Include labels/annotations/owner_kind on each workload
          schema: { type: string, default: 'true' }
        - name: annotation_key
          in: query
          description: Filter by annotation key (fnmatch glob supported)
          schema: { type: string }
        - name: annotation_value
          in: query
          description: Filter by annotation value (fnmatch glob supported; omit for any-value)
          schema: { type: string }
        - name: label_key
          in: query
          schema: { type: string }
        - name: label_value
          in: query
          schema: { type: string }
        - name: owner_name
          in: query
          description: L4-compatibility alias for `workload_name`
          schema: { type: string }
        - name: pod_name
          in: query
          description: Case-insensitive substring match against workload name
          schema: { type: string }
        - name: workload_name
          in: query
          description: Case-insensitive substring match against L7Workload.name
          schema: { type: string }
        - name: filter_noise_annotations
          in: query
          description: Strip Kubernetes infrastructure annotations from response
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: Per-workload inbound/outbound/request/error summary
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /l7/dependencies/tree-summary:
    get:
      tags: [L7 Communications]
      summary: L7 dependency tree with downstream/callers grouped by protocol
      parameters:
        - name: analysis_id
          in: query
          required: true
          schema: { type: string }
        - name: cluster_id
          in: query
          schema: { type: string }
        - name: workload_name
          in: query
          schema: { type: string }
        - name: namespace
          in: query
          schema: { type: string }
        - name: depth
          in: query
          schema: { type: integer, default: 1, minimum: 1, maximum: 3 }
        - name: label_key
          in: query
          schema: { type: string }
        - name: label_value
          in: query
          schema: { type: string }
        - name: annotation_key
          in: query
          schema: { type: string }
        - name: annotation_value
          in: query
          schema: { type: string }
        - name: include_metadata
          in: query
          schema: { type: string, default: 'true' }
        - name: workload_name_exact
          in: query
          description: |
            If true (default) match `workload_name` exactly. Pass false for
            case-insensitive substring match (L4 `owner_name` parity).
          schema: { type: boolean, default: true }
      responses:
        '200':
          description: Tree-structured dependency summary
          content:
            application/json:
              schema: { type: object, additionalProperties: true }

  # ========================================================================
  # L7 Events (Beyla - ClickHouse time-series)
  # ========================================================================
...